refactor: Move Rust tests to separate test files
Moved inline tests from output_blocks.rs and status_formatter.rs to
dedicated test files in rust/tests/ directory for better separation
of concerns and consistency with project conventions.
